% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rippler.R
\name{ripple_download}
\alias{ripple_download}
\title{Pull Data from Ripple Science API}
\usage{
ripple_download(
  url = Sys.getenv("RIPPLER_RIPPLE_URL"),
  study_id = Sys.getenv("RIPPLER_RIPPLE_STUDY_ID"),
  variables = c(standard_vars, custom_vars),
  output_file = NULL
)
}
\arguments{
\item{url}{URL of your Ripple Instance, by default pulled from \code{RIPPLER_RIPPLE_URL} environment variable.}

\item{study_id}{Study ID to pull from, by default pulled from \code{RIPPLER_RIPPLE_STUDY_ID} environment variable.}

\item{variables}{(optional) a vector of variables to pull, by default the complete set of standard and custom variables}

\item{output_file}{(optional) a CSV to write data to}
}
\value{
a data frame containing the data from Ripple Science API (invisibly, if \code{output_file} is specified)
}
\description{
Pull Data from Ripple Science API
}
