% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rippler.R
\name{ripple_upload}
\alias{ripple_upload}
\title{Push data to Ripple Science API}
\usage{
ripple_upload(
  url = Sys.getenv("RIPPLER_RIPPLE_URL"),
  study_id = Sys.getenv("RIPPLER_RIPPLE_STUDY_ID"),
  dat = NULL,
  file = NULL,
  update = c("all", "nocontact", "noexisting")
)
}
\arguments{
\item{url}{URL of your Ripple Instance, by default pulled from \code{RIPPLER_RIPPLE_URL} environment variable.}

\item{study_id}{Study ID to pull from, by default pulled from \code{RIPPLER_RIPPLE_STUDY_ID} environment variable.}

\item{dat}{a data frame to push to Ripple (required if \code{file} not set)}

\item{file}{an xlsx file to push to Ripple (required if \code{dat} is not set)}

\item{update}{update type. One of \code{'all'} (update all participants, the default),
\code{'nocontact'} (do not update any contact info), or
\code{'noexisting'}(do not update any existing participants).}
}
\value{
value returned by \code{httr2::req_perform} on upload
}
\description{
Push data to Ripple Science API
}
